Guard a possibly null pathManager.
If the 'settings-update' event is handled before settings are loaded, `pathManager`
could be null.

Also, extract initializing `pathManager` into a function.

- This happens from both changing a pref (any pref) in the
  first-run window, as well as during normal settings loading.

Note that from first-run we're calling `pathManager.enforce()`,
while from standard startup we're calling `integrationManager.enforce()`
If there's a problem here, it should be covered in a separate issue and PR.

Signed-off-by: Eric Promislow <epromislow@suse.com>