Prevent closing snapshot dialog when lock clears
This aims to ensure that the blocking snapshot dialogs will only close when the dialog window:

  a. receives the `dialog/close` signal
  b. contains a `snapshotEventType` that matches the source of the signal

The original implementation would indiscriminately close a dialog window when the main window receives a signal that a backend lock has cleared. This was problematic in snapshot operations that might result in an error, because they would then clear the backend lock resulting in the blocking dialog closing before the error message could be relayed to the user.

Signed-off-by: Phillip Rak <rak.phillip@gmail.com>