Bundle libcairo inside AppImage
Fedora 39 uses a version of libcairo that is newer than other distros.
This newer version appears to have removed a symbol that is required
for Rancher Desktop. The fix is to bundle the older version of
libcairo inside the AppImage (which is the only package format that
has this problem occurs as of the time of writing).

Signed-off-by: Adam Pickering <adam.pickering@suse.com>