Package: Generate Linux .desktop file from code
In CI (but not locally) the build seems to die when copying over the
.desktop file from the hard-coded file because the destination already
exists.  Work around the issue by generating the file; the functionality
already exists in electron-builder but is normally not called from the zip
build (only the flatpak/snap/fpm builders).  We just have to manually
invoke it.

Signed-off-by: Mark Yen <mark.yen@suse.com>