E2E: Backend: ensure old settings have valid Kubernetes version
If the user previously started Rancher Desktop offline (or otherwise caused
the version to be invalid), we would end up failing when we tried to
compare the existing version to the new proposed version because semver
would fail to parse the empty string.  Avoid this by manually setting the
version to a valid string before continuing.

Signed-off-by: Mark Yen <mark.yen@suse.com>