CI: Package: Disable macOS keychain auto-lock
By default, new keychains are set up to auto-lock after five minutes.  This
happened to be _right_ on the edge of how long everything else takes, so we
got intermittent timeouts during test signing in CI because the system was
waiting for the user to input the (empty) keychain password.

Since we're creating the temporary keychain ourselves anyway, just set it
to never expire instead.

Signed-off-by: Mark Yen <mark.yen@suse.com>