CI: Windows E2E: run on GitHub-hosted runners
GitHub-hosted runners now allow nested virtualization, so we can just use
those instead of trying to use our own.  This has the advantage that the
runner will always be a fresh VM (so we don't need to worry about having
state from previous runs).  However, this means we need to install WSL2
correctly (it comes with WSL1, but at least that means all the Windows
optional components are installed), and we need to disable Docker Desktop
that is built-in.  Docker Desktop doesn't appear to be installed normally
(there isn't an Add/Remove Programs entry, and no C:\Program Files\Docker)
so it was not possible to properly uninstall it.

Signed-off-by: Mark Yen <mark.yen@suse.com>