Don't do no-op pathManager updates
It is possible that the `update-settings` event from the first-run dialog
arrives before the app can initialize the path manager. In that case we
would end up enforcing, removing, and enforcing again the same manager.

Note that we still seem to be enforcing the same manager 3 times even after
this change, but the additional calls at least don't write to the file.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>