Dependabot: ignore @nuxtjs/eslint-config-typescript >11
Newer versions try to auto-detect for Nuxt 3; however, since we don't
actually have Nuxt listed as a dependency, it mis-detects and thinks we are
using Nuxt 3 (using Vue 3) and tries to use the wrong set of lint plugins
for Vue.

Since we're not bumping Nuxt ever, we can use a static set of linters here.

Signed-off-by: Mark Yen <mark.yen@suse.com>