CI: BATS: Allow overriding run ID to download from
This allows an override to which GitHub Actions run id to download from, so
that we can download from a build that isn't the latest completed build.
This is useful to run BATS tests for packaging steps that are still in
progress, as long as the relevant platforms are complete.

Signed-off-by: Mark Yen <mark.yen@suse.com>