CI: Package: Allow skipping signing
When doing a manually-triggered run of packaging, sometimes we don't need
to check signing (and just want the output for some other purpose).  Allow
the signing check steps to be skipped in that case.

Note that this will still be triggered by default, and also it will be run
when it's _not_ manually triggered (e.g. due to a push).

Signed-off-by: Mark Yen <mark.yen@suse.com>