BATS: Windows: Write files with PowerShell
We are having issues with factory reset tests because we write to
setting.json via redirection from the WSL distribution; this seems to
result in the WSL distribution still seeing the file after it has been
deleted from Windows.  Work around this by spawning PowerShell to write the
file instead.

Signed-off-by: Mark Yen <mark.yen@suse.com>