Stop checking for rd-openresty service status
It is not reliable because the service is stopped by background.ts
as soon as the allowed images list is disabled. At this point the
docker daemon and docker context still exist, so the following checks
might still hit the old state, before the backend is restarting.

Where necessary this needs to be controlled by watching the service pid.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>