Deps: lima: simplify extra part parsing
Assume that our releases are always of the correct form, and just parse the
RD-specific number out of the version and compare them as numbers.  This
makes the code a lot shorter, but will be tripped up if there are any non-
conforming releases.

Also modify the main rddepman script so that it would be easier to test
version parsing and comparison locally.

Signed-off-by: Mark Yen <mark.yen@suse.com>