CI: When making the PR to merge to main, use a branch
Since we have things set such that when we merge a PR, GitHub will
automatically delete the head branch, we need to ensure that we create a
disposable branch before we make a PR that merges a release tag back to
`main`.  This happily also means we will have somewhere to push resolutions
to if there are merge conflicts.

Signed-off-by: Mark Yen <mark.yen@suse.com>