Screenshots Kube Backend: Delay creating a client
Creating a Kubernetes client requires a valid kubeconfig (so that we can
figure out the parameters for it); this means that we need to defer
actually doing so until the point where the configuration is ready.  The
screenshot backend (used to fake the list of services) was incorrectly
doing so at the start before actually starting k3s; this lead to things
failing on fresh machines.

Signed-off-by: Mark Yen <mark.yen@suse.com>