E2E: utils: fix startSlowerDesktop()
If we clearly specify the return value as a tuple of two things, rather
than a generic array, we can expose the types of the return value correctly
to avoid casts in the caller.

Additionally, fix startRancherDesktop() so that noModalDialogs defaults to
true instead of false (to match the existing documentation).

Signed-off-by: Mark Yen <mark.yen@suse.com>