BATS: profile/wasm: wait for backend
The next step runs `rdctl` and expects to see a message stating that we're
reconfiguring.  However, if we only wait for the container engine (rather
than waiting for the whole backend to settle), we can end up with the
backend still busy, and end up with a different message about the backend
eventually restarting.  Fix this by waiting for full completion before
changing settings.

Signed-off-by: Mark Yen <mark.yen@suse.com>