Fix lint errors: defer + fatal
When we call `log.Fatalf()` or similar that ends up calling `os.Exit()`, we
don't run any deferred statements.  This gets picked up by the linter as an
issue.  Work around this by wrapping the bulk in an inner function, and
then calling `log.Fatalf()` outside of it (after all of the deferred
statements have been run).

Signed-off-by: Mark Yen <mark.yen@suse.com>