UI: Prefs: Warn on spinkube
If the Kubernetes Spin operator is enabled, show a warning banner if
WebAssembly support is disabled. And do the same on the WebAssembly side in
the same situation as well.

This entailed fixing up the incompatible preferences warning banner to make
it able to navigate across pages (not just tabs).  To help with this, also
expand support in the Vuex transient settings store to natively support
navigation.

Signed-off-by: Mark Yen <mark.yen@suse.com>