WSL: Wait for container engine setup before starting K3s setup
We now install containerd shims (for e.g. wasm) as part of the container
engine setup.  However, the Kubernetes setup needs to wait for that to
complete first so that it can find those shims so that we can build runtime
classes for them.

Explicitly wait for the container engine setup to finish before starting
Kubernetes setup to account for this.  This already happens in the Lima
code path (where fewer things happen in parallel).

Signed-off-by: Mark Yen <mark.yen@suse.com>