BATS: When waiting for moby container engine, wait for the service
On macOS in CI, we can sometimes fail because the docker context isn't
created fast enough.  Add a wait for the service (in the VM) to ensure
we're making forward progress while allowing for a longer wait.  Also
increase the wait delay when trying to get the docker context, for the same
reason.

Signed-off-by: Mark Yen <mark.yen@suse.com>