Lima: Drop Kubernetes endpoint from docker context
This appears to have been dropped in docker CLI v23.0 and doesn't appear to
be used for anything.  Dropping this lets us write out the docker context
faster (because we don't need to wait for Kubernetes to be up before
writing the context out).  This should help with BATS because we can
continue with testing faster.

Signed-off-by: Mark Yen <mark.yen@suse.com>