Moby: WSL: Re-implement file copy to deal with symlinks
Windows doesn't deal with symlinks; re-implement file copy manually to
ensure we dereference them correctly.  We previously used
`tar --dereference` but that flag only works when _creating_ the archive;
it was still failing to handle dangling symlinks.  Reimplement the logic
ourselves so we have better control.

This fixes the BATS test extensions/containers.

Signed-off-by: Mark Yen <mark.yen@suse.com>