Fixes the wrong use of k8s client-go
Previously we were using the client incorrectly by not passing the
context, this was causing internal network error in the client due to
the API server not being ready. However, we continued and blocked on the
event channel. With the new approach, we can determine when the API server
is ready, this way we return a FATAL error and the process exists. This
allows the init process to retry again until the API server becomes
ready.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>