Extensions: split getting and initializing
At some point we had a call site that needed to both get an extension
manager as well initialize it.  That no longer exists, so simplify the code
by splitting it into two functions.  We also are now documented to return
HTTP 503 if the extension API endpoints are called before the extension
subsystem has been set up.

Signed-off-by: Mark Yen <mark.yen@suse.com>