BATS: Time out if rdctl gets stuck
In CI, it seems like BATS can get stuck waiting for the backend to be
ready; that is, we end up waiting for the backend to be started, and all
output ceases until the two hour global timeout.  Try setting a timeout on
the `rdctl` command to see if this will un-jam the tests enough to at least
fail it so we can get more useful logs.

Signed-off-by: Mark Yen <mark.yen@suse.com>