CI: go-work-sync: Do not run for cross-repo PRs
The Actions token provided is unable to commit to forked repositories, even
when the PR is set to allow pushes from maintainers.  Disable this workflow
when dealing with cross-repository pull requests.  We will add a lint to
check for this instead.

Dependabot PRs should still work, as this will be triggered, which will
then allow the PR to pass the lint.

Signed-off-by: Mark Yen <mark.yen@suse.com>