try (in BATS) should separate stdout and stderr
That way we can `run --separate-error try curl …` without using --silent
to get rid of the progress meter.

Ideally we would only re-order stdout/stderr lines when invoked via
`run --separate-stderr`, but I couldn't find a way to make it work.

We should not rely on the relative ordering of mixed filehandle output
anyways, but it may be confusing when we look at logs of failed tests.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>