network-setup: Try to tear down on exit
We seem to have issues in some CI E2E runs where the veth interface in the
default namespace (`veth-rd0`) is not removed and interferes with the next
run.  Attempt to tear it down manually when the process exits, instead of
relying solely on it going away when the namespace is terminated.

This is accomplished by terminating the VM switch process when the
network-setup process exits (via the use of contexts).

Signed-off-by: Mark Yen <mark.yen@suse.com>