Wait for traefik before fetching it's IP
On Windows get_host() gets the IP address from the traefik service,
and will fail with an empty return value if the service is not yet
ready.

The assert failures in get_host() also don't cause a direct test failure
if $(get_host) is just interpolated into a string, so it needs to be
called by itself.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>