Fix RD_USE_NETWORKING_TUNNEL=false
The old code relied on the default being false, so only passed the option
to the app when RD_USE_NETWORKING_TUNNEL was true. Since the app default
has changed, this means we could no longer test the legacy mode.

This commit also changes the default for RD_USE_NETWORKING_TUNNEL to true
to match the default app setting.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>