WSL: spawn() via wsl-exec
Executor.spawn() is currently only used by the image scanner.  That expects
to be able to use the docker credential helper, so it needs to be run from
within the network namespace (to be able to contact the front end).  So we
can change Executor.spawn() to always run within the network namespace.

Signed-off-by: Mark Yen <mark.yen@suse.com>