Dashboard: Fix websocket requests
The http-proxy-middleware bump missed the fact that now websocket requests
end up with the proxied path in front of the client path, resulting in the
server seeing requests like `/k8s/k8s/clusters/local/v1/subscribe`.  This
breaks the request.  However, we can't just set `prependPath` to `false` in
the options, as that breaks non-websocket requests instead.  So we need to
manually strip the duplicate prefix instead.

Signed-off-by: Mark Yen <mark.yen@suse.com>
Fixes: 37744253c488a7f6cb02ba68c5a3482369d13ab9