CI: go.work sync: use `pull_request_target`
We seem to be having issues accessing secrets in the workflow even though
we only process pull requests from the same repository (where documentation
claims secrets should be available).  Switch to using `pull_request_target`
to try to get access to those secrets.

Since we still only run this job if the pull request is from the same
repository (with an explicit manual check), this should still be safe as it
only applies to people with write access to the repository.

Signed-off-by: Mark Yen <mark.yen@suse.com>