CI: docker-cli-monitor: Be more strict on matching existing issues
Previously, we looked for any issue that contained the script name when
considering what to close.  That pulled in issues _about_ the script,
instead of issues _created by_ the script.  Update the queries to be more
strict (by including more of the generated title) so that we don't touch
issues referring to the script.

Signed-off-by: Mark Yen <mark.yen@suse.com>