BATS: try to auto-detect Rancher chart version
The Rancher Helm chart might not support the requested Kubernetes version.
If a specific version was not specified, try to iterate through the
available chart versions to find the minimum chart version that is
compatible with the current Kubernetes version which is also at least the
default version (currently 2.7.0).  Note that newer Kubernetes versions do
not currently have any usable Rancher chart versions at this point.

If we determine that a chart version / Kubernetes version combination does
not work, we skip all tests using that version after making sure it is
marked as failed.

To support newer Rancher chart versions, also change the string we are
looking for in the expected output as we no longer see the literal string
"Rancher Dashboard" in the initial response.

Signed-off-by: Mark Yen <mark.yen@suse.com>