BATS: k8s/helm-install-rancher: Look for incremental progress
This test is prone to failing on macOS CI, possibly because the runners are
somewhat slower.  Try to improve this by _not_ waiting for the helm chart
deployment to finish, but instead manually checking for key log lines in
the containers until they are actually ready.

This also does a couple other things to help this test pass:
- Pre-pull the rancher image to ensure the image pulling doesn't make the
  machine busier than necessary.
- Update the arguments for the cert-manager chart (removing deprecated
  usage).

Signed-off-by: Mark Yen <mark.yen@suse.com>