BATS: extensions: Use exe on Windows
Electron bump to 30.1.1 picked up NodeJS 20.14.0 (previously 20.11.1),
which contains a fix for CVE-2024-27980 and no longer allows running *.cmd
on Windows without using `{ shell: true }`.  We were using that as a script
to run in the BATS and E2E tests for extensions.  Work around the issue by
building a proper golang executable that does the same thing.

Signed-off-by: Mark Yen <mark.yen@suse.com>