CI: BATS: Fix defaulting to tunneled networking on Windows
When the workflow is run on a schedule, none of the inputs are set; in that
case, `inputs.rd-use-networking-tunnel` is blank rather than `true` or
`false`.  In order to make the comparison, we need to convert the input to
a string first, then check if it equals the literal string `false` in order
to correctly default to true.

Signed-off-by: Mark Yen <mark.yen@suse.com>