Add diagnostic to verify kubeConfig
When WSL integration is enabled we would need to check the
content of the kubeConfig to see if it includes Rancher Desktop
only config or not, prior to creating the symlink. Now, we can
trigger a diagnostic that prompts the user to manually create the
symlink if we encounter non-rancher desktop configuration.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>