src/go: Make everything buildable on all platforms
This ensures that all go packages have at least some file, and files that
are needed to be platform-specific are marked as such.  This is needed to
ensure that we can run `golanci-lint` on all the packages without erroring
out.

Note that while things build, they do not do any meaningful work.

Signed-off-by: Mark Yen <mark.yen@suse.com>