BATS: Fix picking k3s version
We were, confusingly, using RD_KUBERNETES_PREV_VERSION in most places to
pick the Kubernetes version to test.  Switch to using the more sensible
RD_KUBERNETES_VERSION instead.  RD_KUBERNETES_PREV_VERSION is now
replaced with RD_KUBERNETES_ALT_VERSION, and for the upgrade/downgrade
test we determine which of the two is lower and do the test accordingly.

Signed-off-by: Mark Yen <mark.yen@suse.com>