Avoid importing `k3sHelper` from UI
Because `k3sHelper.ts` pulls in `@pkg/utils/paths` (which is main-process
only and spawns `rdctl`), we must not use it from the renderer process as
that will cause an error when the relevant window opens.  Instead, move the
Kubernetes version helpers into a new file that does not import it.

While we're at it, avoid sending `semver.SemVer` objects across processes
because that does not work anyway.  See #7208 (though we still need to do
an audit to make sure there are no other non-cloneable objects being
passed).

Signed-off-by: Mark Yen <mark.yen@suse.com>