Integrations: Update macOS/Linux docker CLI plugins path
The previous commit downloaded the plugins to a different directory; update
the macOS and Linux integration manager to use that directory instead.
Windows will be fixed later.  We still add symlinks to `~/.rd/bin/` so that
if the user runs e.g. `docker-compose` instead of `docker compose`, the
executable will still be found.

Signed-off-by: Mark Yen <mark.yen@suse.com>