Integrations: Windows: Update docker CLI plugins
Instead of copying the docker CLI plugins to `~/.docker/cli-plugins`, set
up `~/.docker/config.json` to add the Rancher Desktop docker CLI plugins
directory as an extra directory to scan for plugins.  On Windows, this
appears to take precedence over `~/.docker/cli-plugins`, therefore
resolving issues with not updating the plugins.  For symmetry, also do the
same for WSL distributions.

Signed-off-by: Mark Yen <mark.yen@suse.com>