Remove enableListeners from APITracker
The original rationale for including the `enableListeners` flag in the `APITracker` was
to address specific Kubernetes versions that required a port forwarding fix. The need
for this fix was determined based on the Kubernetes version as follows:

- **Versions < 1.21**: No fix required.
- **Version 1.21**: Fix needed if the patch version is 12 or higher.
- **Version 1.22**: Fix needed if the patch version is 10 or higher.
- **Version 1.23**: Fix needed if the patch version is 7 or higher.
- **Versions >= 1.24**: Fix required.
- **Versions with a major version number other than 1**: Fix required.

Initially, the approach was to create local network listeners for these specific Kubernetes
versions to leverage the WSL listener forwarding feature. However, since we no longer depend
on WSL’s listener forwarding, we can now directly establish connections to the container
using the NodePort entries in iptables. As a result, the `enableListeners` flag is no longer
necessary for handling port forwarding fixes, and connections can be managed more directly
through iptables configurations. Therefore, calling the host switch API
should be adequate for all k8s services that either require the fix or
not.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>