Fix fallback when no Kubernetes versions are available
If Rancher Desktop was previously started with a version, but the user
wipes the `k3s-versions.json` cache and we fail to download a new copy,
handle the error and disable Kubernetes instead of throwing an error.  To
make sure the user is aware of the issue, also pop a diagnostic to describe
the situation.

Signed-off-by: Mark Yen <mark.yen@suse.com>