Diagnostics: bin-in-shell: Fix existing PATH
If Rancher Desktop was launched with the PATH such that `~/.rd/bin/` is
already in the PATH (e.g. because integration was previously set up when we
logged in), we would end up always passing the test even when the
configuration was set to disable it (which would have otherwise been
reflected the next time the user logs in).  Fix this by explicitly cleaning
the PATH before we spawn the shell to do the check.

This also fixes an error where if a shell was not installed we throw an
exception (that gets logged but does not display in the UI) instead of
passing the test.  This also means if the user installs a missing shell
while Rancher Desktop was running, re-running the check will detect that.

Signed-off-by: Mark Yen <mark.yen@suse.com>