Bind WSL-Proxy Process to Specified IP Address
This changes the binding of the WSL-Proxy process from always
using localhost to utilizing the specified host IP address for
port bindings.

Previously, binding exclusively to localhost caused issues for
users who needed to define port bindings that extended beyond
this address. The new approach ensures that the proxy listens
on the provided hostIP.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>