Bump google/go-github/v55 → v66
Since they change the package name for every release this package
isn't bumped by dependabot. Which means we were left with a bunch
of indirect dependencies with known vulnerabilities. Which doesn't
matter that much because we don't actually use the github-runner
monitor anymore. But they show up in our code scanning results.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>