Point iptables chainName function to upstream
Previously, due to limited Go version support, we had to maintain
a custom implementation of the function that containerd uses to
generate iptables chain names. With the Go version bump, we can now
point this function back to the upstream implementation.

Additionally, this change updates the code to output the error message
from running the iptables command, providing better debugging information
compared to just the error code.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>