*nix: Kill process group when main process exits
This adds a `rdctl internal process wait-kill` command which waits for the
given pid to exit, and then kills its whole process group.  This is used to
kill any extension-spawned processes (as long as they haven't detached into
a new process group).  This is only needed on linux / darwin because
Windows has a different implementation where extension-spawned processes
are moved into a job that terminates when the main process terminates
instead.

Signed-off-by: Mark Yen <mark.yen@suse.com>