Pin bundled spin version incl plugins and templates
Due to recent backwards incompatible changes in the spinkube operator
ecosystem we need to tightly control the version of the spin executable
and kube plugin to ensure they are compatible with the spinkube operator.

We use a shim script (or executable on Windows) to set SPIN_DATA_DIR
to $APPDATA/rancher-desktop-spin and reinstall plugins and templates
to the pinned versions on each start of the application (if Wasm support
is enabled).

As usual, users can put a spin version managed by themselves on the
front of the PATH if they want to use a different version.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>