Create sudoers file internally instead of calling `limactl sudoers`
The `limactl sudoers` command will fail if socket_vmnet is not yet
installed into the secure path. So we would need to ask the user for
the sudo password twice: once to install socket_vmnet, and then again
to install the sudoers file. By "guessing" the content beforehand we
can install both at the same time.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>