rdctl: Don't use process groups on Linux
Disable killing all processes in the process on Linux, as Electron does not
create a new process group when running from the RPM/deb package.  Leave it
using process groups on macOS, as launchd still does it.

It is expected that we will revert this once we sort out process groups on
Linux.

Signed-off-by: Mark Yen <mark.yen@suse.com>