Linux: Set process group
On Linux, manually set the process group for the main Electron process to
its pid so that we can use it to terminate any processes we spawn as part
of extensions.

Note that this happens after Electron has started other processes; hence
some of the Electron-internal processes will have the old process group.
This is not an issue as they will go away once the main process does.

Signed-off-by: Mark Yen <mark.yen@suse.com>