Work around types in `@types/tar-stream`
The bump of TypeScript & `@types/node` introduced new requirements for all
`AsyncIterator` objects to implement `AsyncIteratorObject` which must in
turn implement `AsyncDisposable`.  `@types/tar-stream` does not seem to
handle this yet.  (This is only a typing issue, and does not affect the
actual library.)

This is essentially the `@types/tar-stream` version of
https://github.com/DefinitelyTyped/DefinitelyTyped/discussions/71307

Signed-off-by: Mark Yen <mark.yen@suse.com>