Manually downgrade github.com/lima-vm/lima
Not manually downgrading lima was causing the gvisor-tap-vsock to revert back to the latest version 0.8.1.
Also, the gvisor-tap-vsock was reverted back to version 0.7.5 for Networking and guestAgent project.
As part of the 1.17.0 release, we upgraded the gvisor-tap-vsock library to version v0.8.1.
However, this version has caused issues with DNS resolution behind VPNs and other name resolution problems.
To address these issues, we are planning to release a patch, and this downgrade is part of that solution.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>