WiX: Error out if WSL is not installed
The WSL installer has changed again, such that it always uses an MSI
installer, so we cannot install it from within our installer.  Rather than
trying to work around this, just detect the situation and abort the install
with a message so the user can manage it separately.  This ensures that we
do not have to detect network state, etc.

If we wish to have the installer manage WSL installation in the future, it
would be better to create an exe installer that bundles the MSI, and from
there launch the WSL installer separately (outside of the Windows Installer
database lock).

Signed-off-by: Mark Yen <mark.yen@suse.com>