Fix moby OpenAPI generation
The OpenAPI file now contains an example that is too large to fit into an
int64 but does fit in an uint64; however, our YAML parser unconditionally
parses it as an int64.  Work around the issue by manually doing a string
replacement on the file before passing it to the OpenAPI parser.

Signed-off-by: Mark Yen <mark.yen@suse.com>