RDX: Fake CORS response
When running extensions, we don't really want to be doing CORS (since the
extensions are presumably trusted and should be able to access anything on
the internet).  Disabling `webSecurity` broke the preload script for
unknown reasons; so we fake it as a workaround.

Signed-off-by: Mark Yen <mark.yen@suse.com>
(cherry picked from commit 39304822d7f25071db362f1c3b63e21f8629475c)