nerdctl-stub: Allow positional arguments mixed with flags
Sometimes it would be convienent for the user to mix positional arguments
with flags, e.g. `nerdctl build --tag foo . --file Containerfile.custom`.
Support this case by parsing all flags, then dumping all positional
arguments later.  This can only work if no command contains both positional
arguments as well as subcommands (so that we can reliably detect them); add
a case in the help output parser to check this.

This also adds a `yarn generate:nerdctl-stub --verbose` that gets passed to
the generator so it is easier to see what it is doing.

Signed-off-by: Mark Yen <mark.yen@suse.com>