RDX: UI: Move extension upgrade check to vuex store
This moves calculating whether an extension can be upgraded into the vuex
store instead of doing it from the UI.  This lets us centralize some code
duplication, and preps us for using it elsewhere.

This also renames the `extensions/list` getter to
`extensions/installedExtensions` so it reads better at usage sites.

Signed-off-by: Mark Yen <mark.yen@suse.com>