BATS: Tag tests that pass on experimental openSUSE distro
Currently only some of the BATS tests pass on the experimental openSUSE
distro; this tags them so they can be selected with:

  ./bats-core/bin/bats --filter-tags opensuse tests/*/

This currently only works correctly on macOS; other platforms will need to
be tested separately.

Note that for `RD_CONTAINER_ENGINE=moby` the docker CLI plugins need to be
correctly set up first.

Signed-off-by: Mark Yen <mark.yen@suse.com>