Fix slowly growing dialogs on Windows
On Windows, if the DPI changed during RD's run, or (possibly) if the
primary display has different DPI settings than the display the dialog is
shown on, we end up with a window that keeps growing taller to infinity.
Fix this by checking (only on Windows) the body height compared to the
document element height, and only resize if the body is shorter.

Signed-off-by: Mark Yen <mark.yen@suse.com>