Fix too-small first-run dialog on Windows
It appears that for the first-run dialog, the document height ends up
being high enough but the actual window (content) height is too small.
Adjust the preferred-size-changed check to check if the content height
is at least the body height (instead of checking the document height
against the body height), to ensure that we can always display the body
fully.

Fixes: 31564297ef3d093dbb3bfe36f405b1e14651c24d
Signed-off-by: Mark Yen <mark.yen@suse.com>