Bump minimum supported Kubernetes version to 1.25.3
This is necessary because containerd 2.0 no longer support the CRI v1alpha2
API.

We needed to move at least to 1.23+, but that is also ancient. Moving to
1.25+, which is the oldest version that still gets extended support on EKS
and is also the oldest version supported by Rancher 2.8 (which is also
already end-of-life).

Move to 1.25.3+ so I could delete some special case code needing to enable
IPv4 ip forwarding to allow traefik to work on WSL2.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>