extensions: set PATH when spawning host executables
This ensures that the platform's bin directory (containing, among other
things, docker credential helpers) are added to PATH when we spawn host
executables.  This is needed if the extension bundles tools like `helm` or
`kubectl`.

Note that we add them to the end so the extension can still override it.
Also, we add the rest of the environment variables too, so that for example
`kubectl` can find `$HOME`.

Signed-off-by: Mark Yen <mark.yen@suse.com>