screenshots: mac: don't capture attached windows
Since we now set the main window as the parent of the preferences window,
by default `screencapture` will also include the main window when we
attempt to take a screenshot of the preferences window.  Add the `-a`
option to `screencapture` (_Do not capture attached windows._) to avoid
this behaviour and only include the preferences window itself.

Signed-off-by: Mark Yen <mark.yen@suse.com>