Settings: Make VZ default on macOS >= Ventura
We technically don't support macOS older than Ventura, but doing the check
means unsupported people don't get an ugly error on start.

The E2E tests changes are needed because we flip the mount to 9p which is
not support on VZ, so when we clicked on "VZ" we ended up clicking on a
link to the mount page instead.  Providing explicit positions avoids the
issue.

Signed-off-by: Mark Yen <mark.yen@suse.com>