ESLint: Fully detach from dashboard
This migrates our ESLint configuration to the new style, and fully drops
the configuration from rancher/dashboard which we have diverged from
anyway.

Note that a significant subset of the rules are being ignored right now
because we're actually linting more files; this will be fixed in a
follow-up commit.

Signed-off-by: Mark Yen <mark.yen@suse.com>