K3sHelper: uninstalling helm charts: accept 429 as needing retry
When we get a HTTP 429 when uninstalling helm charts, accept that as an
error that requires retrying in addition to 503.  Also fix up the message
due to differences in the error shape.

Also, when retrying GitHub API requests, try to obey the HTTP headers that
describe how long to wait until the API limits get reset.

Signed-off-by: Mark Yen <mark.yen@suse.com>