Vuex: Implement container-engine store
This creates a new vuex store that maintains the namespaces, containers,
and volumes state.  This lets us use it for the container and volume pages.

This lets us share some code between the two pages, and puts the things we
get from the API into a store so it can be debugged more easily using the
Vue dev tools.

Signed-off-by: Mark Yen <mark.yen@suse.com>