BATS: registry: Work around containerd failing to restart containers
When containerd is restarted, sometimes it seems to fail to correctly
restart containers (that are set to --restart=always).  Instead, the
container is wedged in the "created" state and attempting to start it
results in an error.  It is possible to work around this by restarting
containerd until that container is no longer in the "created" state.

Signed-off-by: Mark Yen <mark.yen@suse.com>