lima: Default to SSH port forwarder
Change the default to use the SSH port forwarder instead of the GRPC one
due to reported issues.  Also force set the environment variable so we can
ignore Lima defaults.

There has been issues intermittently reported upstream that makes us not
confident enough in the GRPC stack to set it as the default.  We haven't
released anything with the defaults flipped yet, so migration should not be
needed at this point.

Signed-off-by: Mark Yen <mark.yen@suse.com>