UI: Volumes: Abort list on unmount
In the volume browsing page, if we unmount, terminate the process we use to
list files in a volume early.  This is needed because that process can take
a while to run, and in the E2E case it races against trying to delete the
volume.  We often lose that race on Linux, and the test fails because the
volume could not be deleted because it is still in use by the container
trying to list files.

Signed-off-by: Mark Yen <mark.yen@suse.com>