UI: SortableTable: fix row key
Instead of using the row number as the key, use the row key field.  This
fixes issues where a row gets hidden and the next row takes its place,
causing the vue component state to be erroneously kept.

This is already fixed (differently) in @rancher/components and we'll pick
up their fix when we next update.

Signed-off-by: Mark Yen <mark.yen@suse.com>